Fix reconnect
If the watchdog trigger connection_lost and the serial connection if already
open, it is not possible to close/abort the connection.
Old:
- Watchdog timout
- connection_lost (transport set to None)
- _reconnect
  - _disconnect (which could not abort, because transport not set anymore)
  - _create_connection
    Call pyserial-asyncio: create_serial_connection -> _ensure_reader -> _read_ready -> _close -> connection_lost (end here, because of lock)

New workflow could abort the transport and the new connection will work.
Anyway connection_lost will be called twice:
1. by the watchdog (with TimeoutError)
2. by _reconnect -> _disconnect -> abort
The last one will do nothing more, because of the lock of _reconnect